fix file upload bug

+++ models/oldmodel/billinginfo.js
@@ -4,8 +4,8 @@ module.exports = (sequelize, DataTypes) => {
     visitId: DataTypes.INTEGER,
     billingCat: DataTypes.INTEGER,
     billingSubCat: DataTypes.INTEGER,
-    value: DataTypes.FLOAT,
-    approved: DataTypes.FLOAT
+    value: DataTypes.BIGINT,
+    approved: DataTypes.BIGINT
   }, {});
   BillingInfo.associate = function(models) {
     // associations can be defined here
